change implement

+++ src/main/java/edu/icet/repository/StudentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface StudentRepository extends JpaRepository<StudentEntity,Integer> {
+public interface StudentRepository extends JpaRepository<StudentEntity,Long> {
 
 
 }
